+++ src/providers/clearbit.js
@@ -1,18 +1,17 @@
 'use strict'
 
+const { stringify } = require('querystring')
 const got = require('got')
 
-const { avatarSize } = require('../constant')
+const { AVATAR_SIZE } = require('../constant')
 
 module.exports = async url => {
-  const logoUrl = `https://logo.clearbit.com/${url}?size=${avatarSize}`
+  const logoUrl = `https://logo.clearbit.com/${url}?${stringify({
+    size: AVATAR_SIZE
+  })}`
 
-  try {
-    await got.head(logoUrl)
-    return logoUrl
-  } catch (err) {
-    return null
-  }
+  await got.head(logoUrl)
+  return logoUrl
 }
 
 module.exports.supported = {