+++ src/providers/github.js
@@ -1,16 +1,16 @@
 'use strict'
 
-const { avatarSize } = require('../constant')
+const { stringify } = require('querystring')
 const got = require('got')
 
+const { AVATAR_SIZE } = require('../constant')
+
 module.exports = async username => {
-  try {
-    const logoUrl = `https://github.com/${username}.png?size=${avatarSize}`
-    await got.head(logoUrl)
-    return logoUrl
-  } catch (err) {
-    return null
-  }
+  const logoUrl = `https://github.com/${username}.png?${stringify({
+    size: AVATAR_SIZE
+  })}`
+  await got.head(logoUrl)
+  return logoUrl
 }
 
 module.exports.supported = {