+++ src/providers/gravatar.js
@@ -1,8 +1,10 @@
 'use strict'
 
+const { stringify } = require('querystring')
 const crypto = require('crypto')
+const got = require('got')
 
-const { avatarSize } = require('../constant')
+const { AVATAR_SIZE } = require('../constant')
 
 const md5 = str =>
   crypto
@@ -10,10 +12,14 @@ const md5 = str =>
     .update(str)
     .digest('hex')
 
-module.exports = async (username, fallback) =>
-  `https://gravatar.com/avatar/${md5(
-    username
-  )}?size=${avatarSize}&d=${fallback}`
+module.exports = async username => {
+  const avatarUrl = `https://gravatar.com/avatar/${md5(username.trim().toLowerCase())}?${stringify({
+    size: AVATAR_SIZE,
+    d: '404'
+  })}`
+  await got.head(avatarUrl)
+  return avatarUrl
+}
 
 module.exports.supported = {
   email: true,