+++ src/providers/youtube.js
@@ -1,50 +1,34 @@
 'use strict'
 
-const { chain } = require('lodash')
-const request = require('request-promise')
-const url = require('url')
-const constant = require('../constant')
+const { stringify } = require('querystring')
+const { isNil, get } = require('lodash')
+const pAny = require('p-any')
+const got = require('got')
 
-async function getChannelInfo (channelUrl) {
-  const parts = url.parse(channelUrl).pathname.split('/')
+const { YOUTUBE_API_KEY } = require('../constant')
 
-  const slug = chain(parts)
-    .filter(p => p !== '')
-    .last()
-    .value()
+const getAvatarUrl = async (username, bySlugProp) => {
+  const apiUrl = `https://content.googleapis.com/youtube/v3/channels?${stringify({
+    part: 'id,snippet',
+    [bySlugProp]: username,
+    key: YOUTUBE_API_KEY
+  })}`
 
-  const options = {
-    method: 'GET',
-    url: 'https://content.googleapis.com/youtube/v3/channels',
-    qs: {
-      part: 'id,snippet',
-      key: constant.youtubeApiKey
-    },
-    headers: { 'cache-control': 'no-cache' },
-    json: true
-  }
-  if (channelUrl.includes('/user')) {
-    options.qs.forUsername = slug
-  } else {
-    options.qs.id = slug
+  const { body } = await got(apiUrl, { responseType: 'json' })
+  const avatarUrl = get(body, 'items[0].snippet.thumbnails.medium.url')
+
+  if (isNil(avatarUrl)) {
+    throw new Error(`YouTube avatar not detected for '${bySlugProp}'`)
   }
 
-  const body = await request(options)
-  if (body && body.pageInfo && body.pageInfo.totalResults > 0) {
-    return body.items[0]
-  }
+  return avatarUrl
 }
 
-module.exports = async username => {
-  const youtubeChannelInfo = await getChannelInfo(username)
-  return (
-    youtubeChannelInfo.snippet &&
-    youtubeChannelInfo.snippet.thumbnails.default.url
-  )
-}
+module.exports = async username =>
+  pAny([getAvatarUrl(username, 'forUsername'), getAvatarUrl(username, 'id')])
 
 module.exports.supported = {
   email: false,
   username: true,
-  domain: true
+  domain: false
 }