+++ CHANGELOG.md
@@ -1,6 +1,284 @@
-# Change Log
+# Changelog
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+
+### [2.2.18](https://github.com/Kikobeats/unavatar/compare/v2.2.17...v2.2.18) (2020-06-07)
+
+### [2.2.17](https://github.com/Kikobeats/unavatar/compare/v2.2.16...v2.2.17) (2020-05-28)
+
+### [2.2.16](https://github.com/Kikobeats/unavatar/compare/v2.2.15...v2.2.16) (2020-05-16)
+
+### [2.2.15](https://github.com/Kikobeats/unavatar/compare/v2.2.14...v2.2.15) (2020-05-12)
+
+### [2.2.14](https://github.com/Kikobeats/unavatar/compare/v2.2.13...v2.2.14) (2020-05-04)
+
+### [2.2.13](https://github.com/Kikobeats/unavatar/compare/v2.2.12...v2.2.13) (2020-03-28)
+
+### [2.2.12](https://github.com/Kikobeats/unavatar/compare/v2.2.11...v2.2.12) (2020-03-25)
+
+### [2.2.11](https://github.com/Kikobeats/unavatar/compare/v2.2.10...v2.2.11) (2020-02-04)
+
+### [2.2.10](https://github.com/Kikobeats/unavatar/compare/v2.2.9...v2.2.10) (2020-02-03)
+
+### [2.2.9](https://github.com/Kikobeats/unavatar/compare/v2.2.8...v2.2.9) (2020-01-31)
+
+### [2.2.8](https://github.com/Kikobeats/unavatar/compare/v2.2.7...v2.2.8) (2020-01-28)
+
+### [2.2.7](https://github.com/Kikobeats/unavatar/compare/v2.2.6...v2.2.7) (2020-01-27)
+
+### [2.2.6](https://github.com/Kikobeats/unavatar/compare/v2.2.5...v2.2.6) (2020-01-27)
+
+
+### Bug Fixes
+
+* gravatar respects default avatar ([fecfebc](https://github.com/Kikobeats/unavatar/commit/fecfebcaa5e02f909de4da8a2cf904e105023adf)), closes [#66](https://github.com/Kikobeats/unavatar/issues/66)
+
+### [2.2.5](https://github.com/Kikobeats/unavatar/compare/v2.2.4...v2.2.5) (2020-01-23)
+
+### [2.2.4](https://github.com/Kikobeats/unavatar/compare/v2.2.3...v2.2.4) (2020-01-20)
+
+### [2.2.3](https://github.com/Kikobeats/unavatar/compare/v2.2.2...v2.2.3) (2020-01-14)
+
+### [2.2.2](https://github.com/Kikobeats/unavatar/compare/v2.2.1...v2.2.2) (2020-01-14)
+
+### [2.2.1](https://github.com/Kikobeats/unavatar/compare/v2.2.0...v2.2.1) (2020-01-14)
+
+## [2.2.0](https://github.com/Kikobeats/unavatar/compare/v2.1.3...v2.2.0) (2019-11-01)
+
+
+### Features
+
+* add gitlab support ([8fc7ed0](https://github.com/Kikobeats/unavatar/commit/8fc7ed07263baed6e9dcf8340511e1bed4c9525c)), closes [#63](https://github.com/Kikobeats/unavatar/issues/63)
+
+### [2.1.3](https://github.com/Kikobeats/unavatar/compare/v2.1.2...v2.1.3) (2019-10-31)
+
+### [2.1.2](https://github.com/Kikobeats/unavatar/compare/v2.1.1...v2.1.2) (2019-10-11)
+
+### [2.1.1](https://github.com/Kikobeats/unavatar/compare/v2.1.0...v2.1.1) (2019-10-10)
+
+## [2.1.0](https://github.com/Kikobeats/unavatar/compare/v2.0.6...v2.1.0) (2019-10-10)
+
+
+### Features
+
+* telegram provider ([#61](https://github.com/Kikobeats/unavatar/issues/61)) ([2332e05](https://github.com/Kikobeats/unavatar/commit/2332e053eaf56b3934d056d09142f1b591877e04))
+
+### [2.0.6](https://github.com/Kikobeats/unavatar/compare/v2.0.5...v2.0.6) (2019-08-20)
+
+### [2.0.5](https://github.com/Kikobeats/unavatar/compare/v2.0.4...v2.0.5) (2019-08-20)
+
+### [2.0.4](https://github.com/Kikobeats/unavatar/compare/v2.0.3...v2.0.4) (2019-08-15)
+
+### [2.0.3](https://github.com/Kikobeats/unavatar/compare/v2.0.2...v2.0.3) (2019-08-15)
+
+### [2.0.2](https://github.com/Kikobeats/unavatar/compare/v2.0.1...v2.0.2) (2019-08-14)
+
+### [2.0.1](https://github.com/Kikobeats/unavatar/compare/v2.0.0...v2.0.1) (2019-08-14)
+
+## [2.0.0](https://github.com/Kikobeats/unavatar/compare/v1.2.28...v2.0.0) (2019-08-11)
+
+### [1.2.28](https://github.com/Kikobeats/unavatar/compare/v1.2.27...v1.2.28) (2019-08-11)
+
+### [1.2.27](https://github.com/Kikobeats/unavatar/compare/v1.2.26...v1.2.27) (2019-06-28)
+
+
+### Build System
+
+* remove github icon ([d4e6970](https://github.com/Kikobeats/unavatar/commit/d4e6970))
+
+
+
+### [1.2.26](https://github.com/Kikobeats/unavatar/compare/v1.2.24...v1.2.26) (2019-06-28)
+
+
+### Bug Fixes
+
+* typo ([80e6590](https://github.com/Kikobeats/unavatar/commit/80e6590))
+
+
+### Build System
+
+* migrate to gulp v4 ([a1e0eb4](https://github.com/Kikobeats/unavatar/commit/a1e0eb4))
+* update travis ([3a792b9](https://github.com/Kikobeats/unavatar/commit/3a792b9))
+
+
+
+### [1.2.25](https://github.com/Kikobeats/unavatar/compare/v1.2.24...v1.2.25) (2019-06-19)
+
+
+### Bug Fixes
+
+* typo ([80e6590](https://github.com/Kikobeats/unavatar/commit/80e6590))
+
+
+### Build System
+
+* update travis ([3a792b9](https://github.com/Kikobeats/unavatar/commit/3a792b9))
+
+
+
+### [1.2.24](https://github.com/Kikobeats/unavatar/compare/v1.2.23...v1.2.24) (2019-06-09)
+
+
+### Bug Fixes
+
+* github icon ([100e34c](https://github.com/Kikobeats/unavatar/commit/100e34c))
+
+
+
+### [1.2.23](https://github.com/Kikobeats/unavatar/compare/v1.2.22...v1.2.23) (2019-06-09)
+
+
+### Bug Fixes
+
+* setup clearbit fallback properly ([da0464d](https://github.com/Kikobeats/unavatar/commit/da0464d))
+
+
+
+### [1.2.22](https://github.com/Kikobeats/unavatar/compare/v1.2.21...v1.2.22) (2019-05-28)
+
+
+
+### [1.2.21](https://github.com/Kikobeats/unavatar/compare/v1.2.20...v1.2.21) (2019-05-20)
+
+
+### Build System
+
+* change git-authors-cli position ([efd9908](https://github.com/Kikobeats/unavatar/commit/efd9908))
+
+
+
+### [1.2.20](https://github.com/Kikobeats/unavatar/compare/v1.2.19...v1.2.20) (2019-05-20)
+
+
+
+### [1.2.19](https://github.com/Kikobeats/unavatar/compare/v1.2.18...v1.2.19) (2019-05-17)
+
+
+
+### [1.2.18](https://github.com/Kikobeats/unavatar/compare/v1.2.17...v1.2.18) (2019-05-13)
+
+
+
+### [1.2.17](https://github.com/Kikobeats/unavatar/compare/v1.2.16...v1.2.17) (2019-05-05)
+
+
+
+## [1.2.16](https://github.com/Kikobeats/unavatar/compare/v1.2.15...v1.2.16) (2019-05-05)
+
+
+
+## [1.2.15](https://github.com/Kikobeats/unavatar/compare/v1.2.14...v1.2.15) (2019-05-03)
+
+
+
+## [1.2.14](https://github.com/Kikobeats/unavatar/compare/v1.2.13...v1.2.14) (2019-04-22)
+
+
+
+## [1.2.13](https://github.com/Kikobeats/unavatar/compare/v1.2.12...v1.2.13) (2019-04-21)
+
+
+
+<a name="1.2.12"></a>
+## [1.2.12](https://github.com/Kikobeats/unavatar/compare/v1.2.11...v1.2.12) (2019-04-06)
+
+
+
+<a name="1.2.11"></a>
+## [1.2.11](https://github.com/Kikobeats/unavatar/compare/v1.2.10...v1.2.11) (2019-04-06)
+
+
+### Bug Fixes
+
+* **package:** update cacheable-response to version 1.4.0 ([#45](https://github.com/Kikobeats/unavatar/issues/45)) ([65ede42](https://github.com/Kikobeats/unavatar/commit/65ede42))
+
+
+
+<a name="1.2.10"></a>
+## [1.2.10](https://github.com/Kikobeats/unavatar/compare/v1.2.9...v1.2.10) (2019-04-06)
+
+
+
+<a name="1.2.9"></a>
+## [1.2.9](https://github.com/Kikobeats/unavatar/compare/v1.2.8...v1.2.9) (2019-04-04)
+
+
+
+<a name="1.2.8"></a>
+## [1.2.8](https://github.com/Kikobeats/unavatar/compare/v1.2.7...v1.2.8) (2019-03-13)
+
+
+### Bug Fixes
+
+* **package:** update p-timeout to version 3.0.0 ([#38](https://github.com/Kikobeats/unavatar/issues/38)) ([7201b1e](https://github.com/Kikobeats/unavatar/commit/7201b1e))
+
+
+
+<a name="1.2.7"></a>
+## [1.2.7](https://github.com/Kikobeats/unavatar/compare/v1.2.6...v1.2.7) (2019-03-10)
+
+
+### Bug Fixes
+
+* **package:** update jsdom to version 14.0.0 ([#36](https://github.com/Kikobeats/unavatar/issues/36)) ([95f02bf](https://github.com/Kikobeats/unavatar/commit/95f02bf))
+
+
+
+<a name="1.2.6"></a>
+## [1.2.6](https://github.com/Kikobeats/unavatar/compare/v1.2.5...v1.2.6) (2019-03-04)
+
+
+
+<a name="1.2.5"></a>
+## [1.2.5](https://github.com/Kikobeats/unavatar/compare/v1.2.4...v1.2.5) (2019-02-23)
+
+
+
+<a name="1.2.4"></a>
+## [1.2.4](https://github.com/Kikobeats/unavatar/compare/v1.2.3...v1.2.4) (2019-01-17)
+
+
+
+<a name="1.2.3"></a>
+## [1.2.3](https://github.com/Kikobeats/unavatar/compare/v1.2.2...v1.2.3) (2018-12-19)
+
+
+
+<a name="1.2.2"></a>
+## [1.2.2](https://github.com/Kikobeats/unavatar/compare/v1.2.1...v1.2.2) (2018-12-12)
+
+
+### Bug Fixes
+
+* **package:** update memoize-one to version 4.1.0 ([#27](https://github.com/Kikobeats/unavatar/issues/27)) ([e5d49fa](https://github.com/Kikobeats/unavatar/commit/e5d49fa))
+
+
+
+<a name="1.2.1"></a>
+## [1.2.1](https://github.com/Kikobeats/unavatar/compare/v1.2.0...v1.2.1) (2018-12-11)
+
+
+
+<a name="1.2.0"></a>
+# [1.2.0](https://github.com/Kikobeats/unavatar/compare/v1.1.0...v1.2.0) (2018-11-07)
+
+
+### Features
+
+* add disable fallback ([9139375](https://github.com/Kikobeats/unavatar/commit/9139375))
+
+
+
+<a name="1.1.0"></a>
+# [1.1.0](https://github.com/Kikobeats/unavatar/compare/v1.0.4...v1.1.0) (2018-11-07)
+
+
+### Features
+
+* add youtube support ([3706ce5](https://github.com/Kikobeats/unavatar/commit/3706ce5)), closes [#13](https://github.com/Kikobeats/unavatar/issues/13)
+
+
 
 <a name="1.0.4"></a>
 ## [1.0.4](https://github.com/Kikobeats/unavatar/compare/v1.0.3...v1.0.4) (2018-11-05)